backend/pyproject.toml
[project]
name = "minidocs-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "html2text>=2025.4.15",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-google-vertexai>=3.0.3",
    "langchain-groq>=1.0.1",
    "markdownify>=1.2.2",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-pptx>=1.0.2",
    "rich>=14.2.0",
]
